Pass Time Processor

+++ src/main/java/banking/DepositCommandProcessor.java
@@ -2,7 +2,7 @@ package banking;
 
 public class DepositCommandProcessor extends CommandProcessor {
     int id;
-    int amount;
+    double amount;
 
     public DepositCommandProcessor(Bank bank) {
         super(bank);
@@ -15,6 +15,6 @@ public class DepositCommandProcessor extends CommandProcessor {
 
     private void getDepositVariables(String[] commandSplit) {
         id = Integer.parseInt(commandSplit[1]);
-        amount = Integer.parseInt(commandSplit[2]);
+        amount = Double.parseDouble(commandSplit[2]);
     }
 }
